requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg2-binary
passlib
pyjwt
